test [nfc]: Move prepareBoringImageHttpClient to test_images.dart
This was already being used for tests outside the test file it was
defined in, and we'd like to start using it in an additional
different test file. So, define it centrally in this place that
makes sense.
