+++ test/test_images.dart
@@ -1,7 +1,25 @@
 import 'dart:async';
 import 'dart:io';
 
+import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
+
+/// Set [debugNetworkImageHttpClientProvider] to return a constant image.
+///
+/// Returns the [FakeImageHttpClient] that handles the requests.
+///
+/// The caller must set [debugNetworkImageHttpClientProvider] back to null
+/// before the end of the test.
+// TODO(upstream) simplify callers by using addTearDown: https://github.com/flutter/flutter/issues/123189
+//   See also: https://github.com/flutter/flutter/issues/121917
+FakeImageHttpClient prepareBoringImageHttpClient() {
+  final httpClient = FakeImageHttpClient();
+  debugNetworkImageHttpClientProvider = () => httpClient;
+  httpClient.request.response
+    ..statusCode = HttpStatus.ok
+    ..content = kSolidBlueAvatar;
+  return httpClient;
+}
 
 class FakeImageHttpClient extends Fake implements HttpClient {
   final FakeImageHttpClientRequest request = FakeImageHttpClientRequest();
