+++ test/widgets/content_test.dart
@@ -1,5 +1,3 @@
-import 'dart:io';
-
 import 'package:checks/checks.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
@@ -26,23 +24,6 @@ import '../test_navigation.dart';
 import 'dialog_checks.dart';
 import 'message_list_checks.dart';
 import 'page_checks.dart';
-
-/// Set [debugNetworkImageHttpClientProvider] to return a constant image.
-///
-/// Returns the [FakeImageHttpClient] that handles the requests.
-///
-/// The caller must set [debugNetworkImageHttpClientProvider] back to null
-/// before the end of the test.
-// TODO(upstream) simplify callers by using addTearDown: https://github.com/flutter/flutter/issues/123189
-//   See also: https://github.com/flutter/flutter/issues/121917
-FakeImageHttpClient prepareBoringImageHttpClient() {
-  final httpClient = FakeImageHttpClient();
-  debugNetworkImageHttpClientProvider = () => httpClient;
-  httpClient.request.response
-    ..statusCode = HttpStatus.ok
-    ..content = kSolidBlueAvatar;
-  return httpClient;
-}
 
 void main() {
   // For testing a new content feature:
